pyproject.toml
[project]
name = "beyond_vibes"
version = "0.1.0"
description = "LLM evaluation framework for testing self-hosted, open source models running on llama.cpp."
readme = "README.md"
authors = [
  {name = "Blake Hamm", email = "blake.j.hamm@gmail.com"},
]
license = {text = "BSD license"}
requires-python = ">=3.12"
dependencies = [
  "typer",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "pyyaml",
  "minio",
  "huggingface-hub",
  "python-dotenv",
]

[project.scripts]
beyond-vibes = "beyond_vibes.cli:app"

[project.optional-dependencies]
dev = [
  "cookiecutter>=2.6.0",
  "cruft>=2.15.0",
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "coverage>=7.6.1",
  "ruff>=0.5.7",
  "pre-commit>=3.8.0",
  "ipykernel",
  "notebook",
  "jupyter",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff.lint]
select = ["I", "F", "E", "W", "N", "D", "ANN", "B", "A", "T20", "PYI", "Q", "RET", "ARG", "PTH", "PL", "PT"]
ignore = ["D203", "D213"]


[tool.pytest.ini_options]
adopts = "-s"
testpaths = ["tests"]

[tool.coverage] # Coverage config
source = [
  "src",
  "tests",
  "notebooks",
]

[tool.coverage.run]
command_line = "-m pytest"
omit = ["*test*", "*__init__*", ".venv/*"]
include = ["src/*", "tests/*", "notebooks/*"]

# Always include branch coverage
branch = true

[tool.coverage.report]
fail_under = 80
ignore_errors = true
show_missing = true
